마이 페이지 조회 기능 관련 타입 변경사항 병합

+++ src/main/java/com/djs/dongjibsabackend/domain/dto/myPage/MyIndicatorResponse.java
@@ -10,10 +10,10 @@ import lombok.NoArgsConstructor;
 public class MyIndicatorResponse {
 
     private double calorieAvg;
-    private Integer sumOfSharingAvailableQty;
+    private double sumOfSharingAvailableQty;
 
     @Builder
-    public MyIndicatorResponse(double calorieAvg, Integer sumOfSharingAvailableQty) {
+    public MyIndicatorResponse(double calorieAvg, double sumOfSharingAvailableQty) {
         this.calorieAvg = calorieAvg;
         this.sumOfSharingAvailableQty = sumOfSharingAvailableQty;
     }
